requirements_pro.txt
# SecureAuth PRO-Level Requirements
# Advanced AI Security System

# === Core Dependencies ===
pymysql==1.1.0
python-dotenv==1.0.0

# === Machine Learning - Core ===
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# === Explainable AI (SHAP) ===
shap==0.42.1

# === Model Persistence ===
joblib==1.3.2

# === Optional: Advanced Features ===
# Uncomment if you want these features:

# IP Geolocation (for country detection)
# geoip2==4.7.0

# Visualization (for SHAP plots)
# matplotlib==3.7.2
# plotly==5.17.0

# Progress bars (nice for training)
# tqdm==4.66.1

# === Security Libraries ===
bcrypt==4.1.2
PyJWT==2.8.0

# === Testing ===
# pytest==7.4.3
# pytest-cov==4.1.0

# === Notes ===
# Install with: pip3 install -r requirements_pro.txt
# Or minimal: pip3 install pymysql numpy pandas scikit-learn shap joblib
